Added RTC support

+++ components/iq2020-dev/iq2020.cpp
@@ -903,28 +903,29 @@ int IQ2020Component::processIQ2020Command() {
 				// Format and publish RTC datetime in ISO 8601 format
 				#ifdef USE_TEXT_SENSOR
 				if (this->rtc_datetime_sensor_) {
-				char datetime_str[20];
-				snprintf(datetime_str, sizeof(datetime_str), "%04d-%02d-%02d %02d:%02d:%02d", 
-						year, month, day, hours, minutes, seconds);
-				this->rtc_datetime_sensor_->publish_state(datetime_str);
+					char datetime_str[20];
+					snprintf(datetime_str, sizeof(datetime_str), "%04d-%02d-%02d %02d:%02d:%02d", 
+							year, month, day, hours, minutes, seconds);
+					this->rtc_datetime_sensor_->publish_state(datetime_str);
 				}
 				#endif
 
-				// Calculate and publish Unix timestamp
+				// Calculate and publish Unix timestamp (64-bit)
 				#ifdef USE_SENSOR
 				if (this->rtc_timestamp_sensor_) {
-				// Simple Unix timestamp calculation
-				// Note: This doesn't account for leap seconds, but is accurate enough for most purposes
-				struct tm timeinfo = {};
-				timeinfo.tm_year = year - 1900;  // tm_year is years since 1900
-				timeinfo.tm_mon = month - 1;     // tm_mon is 0-11
-				timeinfo.tm_mday = day;
-				timeinfo.tm_hour = hours;
-				timeinfo.tm_min = minutes;
-				timeinfo.tm_sec = seconds;
-				timeinfo.tm_isdst = -1;          // Let mktime determine DST
-				time_t timestamp = mktime(&timeinfo);
-				this->rtc_timestamp_sensor_->publish_state((float)timestamp);
+					// Unix timestamp calculation using 64-bit time_t
+					// Note: This doesn't account for leap seconds, but is accurate enough for most purposes
+					struct tm timeinfo = {};
+					timeinfo.tm_year = year - 1900;  // tm_year is years since 1900
+					timeinfo.tm_mon = month - 1;     // tm_mon is 0-11
+					timeinfo.tm_mday = day;
+					timeinfo.tm_hour = hours;
+					timeinfo.tm_min = minutes;
+					timeinfo.tm_sec = seconds;
+					timeinfo.tm_isdst = -1;          // Let mktime determine DST
+					time_t timestamp = mktime(&timeinfo);
+					// Cast to double to preserve full 64-bit precision
+					this->rtc_timestamp_sensor_->publish_state((double)timestamp);
 				}
 				#endif
 			}
